Basic RPC works
To demo running the RPC server, and hitting it with a client, run
python rc_server.py
and in another terminal run
python rc_client.py.
To choose the command that the client request, look at the example from rc_client.py and go from there.

To learn more about grpc for python, and how to got from the .proto files to the generated files look here:
https://developers.google.com/protocol-buffers/docs/pythontutorial

The grpc server and client is not intended to be used as stands here. This is intending as a starting off point.