Grpc and Logging updates
Updated the GRPC function so that results of the automation command are sent to, and displayed by, the client.

Updating the Logging function requirements

Updated the automation.sh script so that it installs proper dependencies, creates ottomate user and assigns ownership of all log files to ottomate. It is expected that the server will run under the user ottomate going forward.