Updating Portainer Invocation
Portainer doesn't add the local docker instance for monitoring unless you explicitly pass it as an argument in the docker-compose invocation. Specifically have to add the flag '-H unix:///var/run/docker.sock' as that is where the docker socket is mounted according to the volumes clause in this docker-compose file.